change SQLite date to integer type

+++ R/TypeInfo.R
@@ -53,7 +53,7 @@ typesR2DBMS <-
                   Oracle = list(double="binary_double", integer="decimal",
                   character="varchar(255)", logical="varchar(5)", date="date"),
                   SQLite = list(double="REAL", integer="INTEGER",
-                  character="TEXT", logical="TEXT", date="REAL"),
+                  character="TEXT", logical="TEXT", date="INTEGER"),
                   EXCEL = list(double="NUMBER", integer="NUMBER",
                   character="VARCHAR(255)", logical="LOGICAL", date="DATE"),
                   DBASE = list(double="Numeric", integer="Numeric",
